implement test for 'Should mint an NFT with the right uri'
